[module-kafka] ContainerMode moved to common package
