+++ cockroachdb/src/test/java/io/goodforgod/testcontainers/extensions/jdbc/CockroachdbLiquibaseMigrationPerMethodTests.java
@@ -2,6 +2,7 @@ package io.goodforgod.testcontainers.extensions.jdbc;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import io.goodforgod.testcontainers.extensions.ContainerMode;
 import org.junit.jupiter.api.*;
 
 @TestcontainersCockroachdb(mode = ContainerMode.PER_CLASS,
