+++ gradle.properties
@@ -1,6 +1,6 @@
 groupId=io.goodforgod
 artifactRootId=testcontainers-extensions
-artifactVersion=0.2.0-SNAPSHOT
+artifactVersion=0.3.0-SNAPSHOT
 
 
 ##### GRADLE #####
