+++ jdbc/src/main/java/io/goodforgod/testcontainers/extensions/jdbc/AbstractTestcontainersJdbcExtension.java
@@ -1,5 +1,6 @@
 package io.goodforgod.testcontainers.extensions.jdbc;
 
+import io.goodforgod.testcontainers.extensions.ContainerMode;
 import java.io.FileWriter;
 import java.io.Writer;
 import java.lang.annotation.Annotation;
