+++ jdbc/src/test/java/io/goodforgod/testcontainers/extensions/jdbc/LiquibaseMigrationPerMethodTests.java
@@ -2,6 +2,7 @@ package io.goodforgod.testcontainers.extensions.jdbc;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import io.goodforgod.testcontainers.extensions.ContainerMode;
 import org.junit.jupiter.api.*;
 
 @TestcontainersJdbc(mode = ContainerMode.PER_CLASS,
