+++ oracle/src/test/java/io/goodforgod/testcontainers/extensions/oracle/OracleContainerPerClassTests.java
@@ -2,7 +2,7 @@ package io.goodforgod.testcontainers.extensions.oracle;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import io.goodforgod.testcontainers.extensions.jdbc.ContainerMode;
+import io.goodforgod.testcontainers.extensions.ContainerMode;
 import io.goodforgod.testcontainers.extensions.jdbc.ContainerOracleConnection;
 import io.goodforgod.testcontainers.extensions.jdbc.JdbcConnection;
 import io.goodforgod.testcontainers.extensions.jdbc.TestcontainersOracle;
